factor main function out of test
